% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophenetic_calc.R
\name{correlationAssessment}
\alias{correlationAssessment}
\title{Calculate the mean Frobenius norm of list of correlation matrices}
\usage{
correlationAssessment(M)
}
\arguments{
\item{M}{list of factor matrices}
}
\value{
average of frobenius norm of correlation matrices (numeric)
}
\description{
Calculate the mean Frobenius norm of list of correlation matrices
}
