% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_V.R
\name{fit_V}
\alias{fit_V}
\title{Update the factor matrix V during the alternating least squares}
\usage{
fit_V(X, W, U, option, formerV = NULL)
}
\arguments{
\item{X}{the matrix to be decomposed (N x T, N is the number of data points, T is the number of features)}

\item{W}{the weight matrix, same size as in X}

\item{U}{learned loading matrix  (N x K, N is the number of data points, K is the number of factors)}

\item{option}{list of all option settings (accesses V, LD, gls,
actively_calibrating_sparsity, posF, regression_method, epsilon,traitSpecificVar,intercept_ubiq)}

\item{formerV}{Optional argument to pass in previous iteration of V as prior to next one.}
}
\value{
List containing the new V ("V") and the sparsity upper limits for each regression steps ("sparsity_space")
}
\description{
Update the factor matrix V during the alternating least squares
}
