% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_autofit_functions.R
\name{runStdPipeClean}
\alias{runStdPipeClean}
\title{Wrapper to run}
\usage{
runStdPipeClean(args, alpha, lambda, opath = "")
}
\arguments{
\item{args}{List of all arguments needed to run the pipeline, including path to input data files}

\item{alpha}{\eqn{\alpha} sparsity parameter for U matrix}

\item{lambda}{\eqn{\lambda} sparsity parameter for V matrix (numeric)}

\item{opath}{(optional) prefix for output files (numeric)}
}
\value{
a list cotaining all return gwasMF
}
\description{
Wrapper to run
}
