% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophenetic_calc.R
\name{scaledCorrelationAssessment}
\alias{scaledCorrelationAssessment}
\title{Various methods to scale and aggregate the magnitude of correlation internal to several matrices}
\usage{
scaledCorrelationAssessment(M, type = "squared")
}
\arguments{
\item{M}{list of matrices to asses (i.e. factor matrix)}

\item{type}{Specify type of scaling to do: options are squared (divides by ncol^2),
average_r2 (takes median of average of covariance per matrix ), or unscaled (no scaling by # cols).
Default is just to scale by # cols in each matrix}
}
\value{
scaled/adjusted correlation structure score
}
\description{
Various methods to scale and aggregate the magnitude of correlation internal to several matrices
}
